Added resposiveness to the graph paper container. Renamed the graph to have a more generalized name.
