+++ client/src/styles/GraphContainer.css
@@ -1,0 +1,5 @@
+.graph-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
